# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CraftedBy API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List all reviews",
                "operationId": "5831a40d4e407d541cb5b9005e8bab5b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by partial comment",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter by rating",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a new review",
                "operationId": "5060970b6e4a82238edce726ac19498d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/reviews/{id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Display a specific review",
                "operationId": "ead1405d310b46f9470d83b00846b5a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update an existing review",
                "operationId": "9a833f183fe00a126e22f75e19e62ccc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review updated successfully"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "operationId": "5cf0071d318478741a5b237c203496f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted successfully"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Review": {
                "required": [
                    "rating",
                    "user_id",
                    "product_id"
                ],
                "properties": {
                    "rating": {
                        "description": "Rating of the product by the user",
                        "type": "integer",
                        "format": "int32",
                        "example": 5
                    },
                    "comment": {
                        "description": "User's comment about the product",
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Great product!"
                    },
                    "user_id": {
                        "description": "UUID of the user who wrote the review",
                        "type": "string",
                        "format": "uuid"
                    },
                    "product_id": {
                        "description": "UUID of the product being reviewed",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "0": "API key login",
                "type": "apiKey",
                "description": "Enter token in format \"Bearer <`token`>\" (mind the space after Bearer)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}